refactor: refactor to work with the new backend API
