+++ tests/orquestra/braket/simulator/localbraket_test.py
@@ -4,10 +4,6 @@
 import numpy as np
 import pytest
 from braket.circuits.noise import Noise
-from orquestra.quantum.api.backend_test import (
-    QuantumSimulatorGatesTest,
-    QuantumSimulatorTests,
-)
 from orquestra.quantum.circuits import CNOT, Circuit, H, X
 from orquestra.quantum.operators import PauliTerm
 
@@ -40,16 +36,16 @@ def noisy_simulator():
 
 
 @pytest.mark.local
-class TestBraketLocalSimulator(QuantumSimulatorTests):
+class TestBraketLocalSimulator:
     def test_setup_basic_simulators(self, wf_simulator):
         assert isinstance(wf_simulator, BraketLocalSimulator)
         # assert wf_simulator.noise_model is None
 
-    def test_run_circuit_and_measure(self):
+    def test_run_and_measure(self):
         # Given
         circuit = Circuit([X(0), CNOT(1, 2)])
         simulator = BraketLocalSimulator()
-        measurements = simulator.run_circuit_and_measure(circuit, n_samples=100)
+        measurements = simulator.run_and_measure(circuit, n_samples=100)
         assert len(measurements.bitstrings) == 100
 
         for measurement in measurements.bitstrings:
@@ -59,20 +55,20 @@ class TestBraketLocalSimulator(QuantumSimulatorTests):
         # Given
         circuit = Circuit([X(0), CNOT(1, 2)], n_qubits=4)
         simulator = BraketLocalSimulator()
-        measurements = simulator.run_circuit_and_measure(circuit, n_samples=100)
+        measurements = simulator.run_and_measure(circuit, n_samples=100)
         assert len(measurements.bitstrings) == 100
 
         for measurement in measurements.bitstrings:
             assert measurement == (1, 0, 0, 0)
 
-    def test_run_circuitset_and_measure(self):
+    def test_run_batch_and_measure(self):
         # Given
         simulator = BraketLocalSimulator()
         circuit = Circuit([X(0), CNOT(1, 2)])
         n_circuits = 5
         n_samples = 100
         # When
-        measurements_set = simulator.run_circuitset_and_measure(
+        measurements_set = simulator.run_batch_and_measure(
             [circuit] * n_circuits, n_samples=[100] * n_circuits
         )
         # Then
@@ -99,25 +95,6 @@ class TestBraketLocalSimulator(QuantumSimulatorTests):
             wavefunction.amplitudes[7], (1 / np.sqrt(2) + 0j), atol=10e-15
         )
 
-    def test_get_exact_expectation_values(self):
-        # Given
-        simulator = BraketLocalSimulator()
-        circuit = Circuit([H(0), CNOT(0, 1), CNOT(1, 2)])
-        qubit_operator = (
-            PauliTerm.identity() * 2
-            - PauliTerm({0: "Z", 1: "Z"})
-            + PauliTerm({0: "X", 2: "X"})
-        )
-        target_values = np.array([2.0, -1.0, 0.0])
-
-        # When
-
-        expectation_values = simulator.get_exact_expectation_values(
-            circuit, qubit_operator
-        )
-        # Then
-        np.testing.assert_array_almost_equal(expectation_values.values, target_values)
-
     def test_get_noisy_exact_expectation_values(self):
         # Given
         noise = 0.0002
@@ -135,15 +112,15 @@ class TestBraketLocalSimulator(QuantumSimulatorTests):
         )
         np.testing.assert_almost_equal(expectation_values.values[1], target_values[1])
 
-    def test_run_circuit_and_measure_seed(self):
+    def test_run_and_measure_seed(self):
         # Given
         circuit = Circuit([X(0), CNOT(1, 2)])
         simulator1 = BraketLocalSimulator()
         simulator2 = BraketLocalSimulator()
 
         # When
-        measurements1 = simulator1.run_circuit_and_measure(circuit, n_samples=1000)
-        measurements2 = simulator2.run_circuit_and_measure(circuit, n_samples=1000)
+        measurements1 = simulator1.run_and_measure(circuit, n_samples=1000)
+        measurements2 = simulator2.run_and_measure(circuit, n_samples=1000)
 
         # Then
         for (meas1, meas2) in zip(measurements1.bitstrings, measurements2.bitstrings):
@@ -165,9 +142,3 @@ class TestBraketLocalSimulator(QuantumSimulatorTests):
 
     def test_get_wavefunction_uses_provided_initial_state(self, wf_simulator):
         pytest.xfail("Braket simulator only accepts zero state as initial state")
-
-
-@pytest.mark.local
-class TestBraketLocalSimulatorGates(QuantumSimulatorGatesTest):
-    atol_wavefunction = 1e-8
-    gates_to_exclude = ["RH"]
